+++ src/api/factory.js
@@ -1,0 +1,7 @@
+import skate from './skate';
+
+export default function (opts) {
+  return function (name) {
+    return skate(name, opts);
+  };
+}